+++ src/muibook/navbar/navbar.js
@@ -221,6 +221,7 @@ class appNavbar extends HTMLElement {
           <app-navbar-link link="/image" title="Image"></app-navbar-link>
           <app-navbar-link link="/media-player" title="Media Player"></app-navbar-link>
           <app-navbar-link link="/model-viewer" title="Model Viewer"></app-navbar-link>
+          <app-navbar-link link="/financial-chart" title="Financial Chart"></app-navbar-link>
           <app-navbar-link link="/video-thumbnail" title="Video Thumbnail"></app-navbar-link>
           <app-navbar-link link="/slide-frame" title="Slide Frame"></app-navbar-link>
         </app-navbar-section>